requirements.txt
anyio==4.8.0
asgiref==3.8.1
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.0.0
fastapi==0.95.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysqlclient==2.1.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.21
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.26.1
typing_extensions==4.12.2
uvicorn==0.15.0
